+++ src/policies/FifoWindowPolicy.cpp
@@ -177,7 +177,7 @@ void FifoWindowPolicy::notifyTouch(Mapping * mapping, size_t index, bool isWrite
 
 	//really do the evict out of the critical section to keep multi-threading
 	//to write data
-	for (size_t i = 0 ; i < cntIdsToEvict; i++)
+	for (int i = 0 ; i < cntIdsToEvict; i++)
 		mappings[i]->evict(this, idsToEvict[i]);
 
 	//notif quota to redistribute if needed
@@ -262,7 +262,7 @@ void FifoWindowPolicy::shrinkMemory(void)
 
 	//really do the evict out of the critical section to keep multi-threading
 	//to write data
-	for (size_t i = 0 ; i < cntIdsToEvict; i++)
+	for (int i = 0 ; i < cntIdsToEvict; i++)
 		mappings[i]->evict(this, idsToEvict[i]);
 }
